Edited Stack to work with doubles and completed RPN

+++ stack/Stack.java
@@ -1,11 +1,11 @@
 package stack;
 
 public class Stack {
-    private int[] elements;
+    private double[] elements;
     private int stackTop;
 
     public Stack(int size) {
-        elements = new int[size];
+        elements = new double[size];
         stackTop = -1;
     }
 
@@ -13,9 +13,9 @@ public class Stack {
         return stackTop<0;
     }
 
-    public void push(int element) {
+    public void push(double element) {
         if (stackTop == elements.length-1) {
-            int[] newElements = new int[elements.length + 1];
+            double[] newElements = new double[elements.length + 1];
             for (int i=0; i<elements.length; i++){
                 newElements[i] = elements[i];
             }
@@ -24,13 +24,13 @@ public class Stack {
         elements[++stackTop] = element;
     }
 
-    public int pop() {
+    public double pop() {
         if (this.isEmpty()) {
             throw new IllegalArgumentException();
         }
         return elements[stackTop--];
         }
-    public int peek(){
+    public double peek(){
         if (this.isEmpty()) {
             throw new IllegalArgumentException();
         }
